rename the last few things
