+++ assets/src/index.ts
@@ -1,1 +1,1 @@
-export { LiveViewOptions, connectAndRun } from "./liveview"
+export { LiveViewOptions, connectAndRun } from "./live_view"
