+++ examples/form/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "axum-liveview": "file:../../assets"
+    "axum-live-view": "file:../../assets"
   },
   "private": true,
   "devDependencies": {
